backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
textblob==0.18.0
nltk==3.9.1
requests==2.32.3
cachetools==5.5.0
apscheduler==3.10.4
pytest==8.3.3
anyio==4.6.0
